putting things together; cleanup
