+++ assets/js/components/TargetChargePlan.story.vue
@@ -1,29 +1,137 @@
 <script setup>
 import TargetChargePlan from "./TargetChargePlan.vue";
 
-const priceSlots1 = [
-	{ price: 0.41, start: "2022-12-22T19:00:00Z" },
-	{ price: 0.22, start: "2022-12-22T20:00:00Z" },
-	{ price: 0.19, start: "2022-12-22T21:00:00Z" },
-	{ price: 0.07, start: "2022-12-22T22:00:00Z" },
-	{ price: 0.33, start: "2022-12-22T23:00:00Z" },
-	{ price: 0.34, start: "2022-12-23T00:00:00Z" },
-	{ price: 0.61, start: "2022-12-23T01:00:00Z" },
-	{ price: 0.57, start: "2022-12-23T02:00:00Z" },
-	{ price: 0.42, start: "2022-12-23T03:00:00Z" },
-];
+const tariff = {
+	rates: [
+		{ start: "2023-01-10T22:00:00+01:00", end: "2023-01-10T23:00:00+01:00", price: 545 },
+		{ start: "2023-01-10T23:00:00+01:00", end: "2023-01-11T00:00:00+01:00", price: 518 },
+		{ start: "2023-01-11T00:00:00+01:00", end: "2023-01-11T01:00:00+01:00", price: 545 },
+		{ start: "2023-01-11T01:00:00+01:00", end: "2023-01-11T02:00:00+01:00", price: 518 },
+		{ start: "2023-01-11T02:00:00+01:00", end: "2023-01-11T03:00:00+01:00", price: 536 },
+		{ start: "2023-01-11T03:00:00+01:00", end: "2023-01-11T04:00:00+01:00", price: 545 },
+		{ start: "2023-01-11T04:00:00+01:00", end: "2023-01-11T05:00:00+01:00", price: 527 },
+		{ start: "2023-01-11T05:00:00+01:00", end: "2023-01-11T06:00:00+01:00", price: 527 },
+		{ start: "2023-01-11T06:00:00+01:00", end: "2023-01-11T07:00:00+01:00", price: 536 },
+		{ start: "2023-01-11T07:00:00+01:00", end: "2023-01-11T08:00:00+01:00", price: 518 },
+		{ start: "2023-01-11T08:00:00+01:00", end: "2023-01-11T09:00:00+01:00", price: 400 },
+		{ start: "2023-01-11T09:00:00+01:00", end: "2023-01-11T10:00:00+01:00", price: 336 },
+		{ start: "2023-01-11T10:00:00+01:00", end: "2023-01-11T11:00:00+01:00", price: 336 },
+		{ start: "2023-01-11T11:00:00+01:00", end: "2023-01-11T12:00:00+01:00", price: 336 },
+		{ start: "2023-01-11T12:00:00+01:00", end: "2023-01-11T13:00:00+01:00", price: 336 },
+		{ start: "2023-01-11T13:00:00+01:00", end: "2023-01-11T14:00:00+01:00", price: 336 },
+		{ start: "2023-01-11T14:00:00+01:00", end: "2023-01-11T15:00:00+01:00", price: 336 },
+		{ start: "2023-01-11T15:00:00+01:00", end: "2023-01-11T16:00:00+01:00", price: 336 },
+		{ start: "2023-01-11T16:00:00+01:00", end: "2023-01-11T17:00:00+01:00", price: 372 },
+		{ start: "2023-01-11T17:00:00+01:00", end: "2023-01-11T18:00:00+01:00", price: 400 },
+		{ start: "2023-01-11T18:00:00+01:00", end: "2023-01-11T19:00:00+01:00", price: 555 },
+		{ start: "2023-01-11T19:00:00+01:00", end: "2023-01-11T20:00:00+01:00", price: 555 },
+		{ start: "2023-01-11T20:00:00+01:00", end: "2023-01-11T21:00:00+01:00", price: 545 },
+		{ start: "2023-01-11T21:00:00+01:00", end: "2023-01-11T22:00:00+01:00", price: 555 },
+		{ start: "2023-01-11T22:00:00+01:00", end: "2023-01-11T23:00:00+01:00", price: 564 },
+		{ start: "2023-01-11T23:00:00+01:00", end: "2023-01-12T00:00:00+01:00", price: 545 },
+		{ start: "2023-01-12T00:00:00+01:00", end: "2023-01-12T01:00:00+01:00", price: 555 },
+		{ start: "2023-01-12T01:00:00+01:00", end: "2023-01-12T02:00:00+01:00", price: 545 },
+		{ start: "2023-01-12T02:00:00+01:00", end: "2023-01-12T03:00:00+01:00", price: 536 },
+		{ start: "2023-01-12T03:00:00+01:00", end: "2023-01-12T04:00:00+01:00", price: 545 },
+		{ start: "2023-01-12T04:00:00+01:00", end: "2023-01-12T05:00:00+01:00", price: 527 },
+		{ start: "2023-01-12T05:00:00+01:00", end: "2023-01-12T06:00:00+01:00", price: 536 },
+		{ start: "2023-01-12T06:00:00+01:00", end: "2023-01-12T07:00:00+01:00", price: 518 },
+		{ start: "2023-01-12T07:00:00+01:00", end: "2023-01-12T08:00:00+01:00", price: 545 },
+		{ start: "2023-01-12T08:00:00+01:00", end: "2023-01-12T09:00:00+01:00", price: 509 },
+		{ start: "2023-01-12T09:00:00+01:00", end: "2023-01-12T10:00:00+01:00", price: 336 },
+		{ start: "2023-01-12T10:00:00+01:00", end: "2023-01-12T11:00:00+01:00", price: 336 },
+		{ start: "2023-01-12T11:00:00+01:00", end: "2023-01-12T12:00:00+01:00", price: 336 },
+		{ start: "2023-01-12T12:00:00+01:00", end: "2023-01-12T13:00:00+01:00", price: 336 },
+		{ start: "2023-01-12T13:00:00+01:00", end: "2023-01-12T14:00:00+01:00", price: 336 },
+		{ start: "2023-01-12T14:00:00+01:00", end: "2023-01-12T15:00:00+01:00", price: 354 },
+		{ start: "2023-01-12T15:00:00+01:00", end: "2023-01-12T16:00:00+01:00", price: 345 },
+		{ start: "2023-01-12T16:00:00+01:00", end: "2023-01-12T17:00:00+01:00", price: 372 },
+		{ start: "2023-01-12T17:00:00+01:00", end: "2023-01-12T18:00:00+01:00", price: 400 },
+		{ start: "2023-01-12T18:00:00+01:00", end: "2023-01-12T19:00:00+01:00", price: 500 },
+		{ start: "2023-01-12T19:00:00+01:00", end: "2023-01-12T20:00:00+01:00", price: 509 },
+		{ start: "2023-01-12T20:00:00+01:00", end: "2023-01-12T21:00:00+01:00", price: 491 },
+		{ start: "2023-01-12T21:00:00+01:00", end: "2023-01-12T22:00:00+01:00", price: 454 },
+		{ start: "2023-01-12T22:00:00+01:00", end: "2023-01-12T23:00:00+01:00", price: 445 },
+		{ start: "2023-01-12T23:00:00+01:00", end: "2023-01-13T00:00:00+01:00", price: 445 },
+		{ start: "2023-01-13T00:00:00+01:00", end: "2023-01-13T01:00:00+01:00", price: 573 },
+		{ start: "2023-01-13T01:00:00+01:00", end: "2023-01-13T02:00:00+01:00", price: 600 },
+		{ start: "2023-01-13T02:00:00+01:00", end: "2023-01-13T03:00:00+01:00", price: 555 },
+		{ start: "2023-01-13T03:00:00+01:00", end: "2023-01-13T04:00:00+01:00", price: 564 },
+		{ start: "2023-01-13T04:00:00+01:00", end: "2023-01-13T05:00:00+01:00", price: 600 },
+		{ start: "2023-01-13T05:00:00+01:00", end: "2023-01-13T06:00:00+01:00", price: 527 },
+		{ start: "2023-01-13T06:00:00+01:00", end: "2023-01-13T07:00:00+01:00", price: 545 },
+		{ start: "2023-01-13T07:00:00+01:00", end: "2023-01-13T08:00:00+01:00", price: 527 },
+		{ start: "2023-01-13T08:00:00+01:00", end: "2023-01-13T09:00:00+01:00", price: 509 },
+		{ start: "2023-01-13T09:00:00+01:00", end: "2023-01-13T10:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T10:00:00+01:00", end: "2023-01-13T11:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T11:00:00+01:00", end: "2023-01-13T12:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T12:00:00+01:00", end: "2023-01-13T13:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T13:00:00+01:00", end: "2023-01-13T14:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T14:00:00+01:00", end: "2023-01-13T15:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T15:00:00+01:00", end: "2023-01-13T16:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T16:00:00+01:00", end: "2023-01-13T17:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T17:00:00+01:00", end: "2023-01-13T18:00:00+01:00", price: 336 },
+		{ start: "2023-01-13T18:00:00+01:00", end: "2023-01-13T19:00:00+01:00", price: 473 },
+		{ start: "2023-01-13T19:00:00+01:00", end: "2023-01-13T20:00:00+01:00", price: 491 },
+		{ start: "2023-01-13T20:00:00+01:00", end: "2023-01-13T21:00:00+01:00", price: 445 },
+		{ start: "2023-01-13T21:00:00+01:00", end: "2023-01-13T22:00:00+01:00", price: 454 },
+		{ start: "2023-01-13T22:00:00+01:00", end: "2023-01-13T23:00:00+01:00", price: 418 },
+		{ start: "2023-01-13T23:00:00+01:00", end: "2023-01-14T00:00:00+01:00", price: 400 },
+		{ start: "2023-01-14T00:00:00+01:00", end: "2023-01-14T01:00:00+01:00", price: 536 },
+		{ start: "2023-01-14T01:00:00+01:00", end: "2023-01-14T02:00:00+01:00", price: 555 },
+		{ start: "2023-01-14T02:00:00+01:00", end: "2023-01-14T03:00:00+01:00", price: 555 },
+		{ start: "2023-01-14T03:00:00+01:00", end: "2023-01-14T04:00:00+01:00", price: 555 },
+		{ start: "2023-01-14T04:00:00+01:00", end: "2023-01-14T05:00:00+01:00", price: 564 },
+		{ start: "2023-01-14T05:00:00+01:00", end: "2023-01-14T06:00:00+01:00", price: 555 },
+		{ start: "2023-01-14T06:00:00+01:00", end: "2023-01-14T07:00:00+01:00", price: 555 },
+		{ start: "2023-01-14T07:00:00+01:00", end: "2023-01-14T08:00:00+01:00", price: 573 },
+		{ start: "2023-01-14T08:00:00+01:00", end: "2023-01-14T09:00:00+01:00", price: 545 },
+		{ start: "2023-01-14T09:00:00+01:00", end: "2023-01-14T10:00:00+01:00", price: 336 },
+		{ start: "2023-01-14T10:00:00+01:00", end: "2023-01-14T11:00:00+01:00", price: 336 },
+		{ start: "2023-01-14T11:00:00+01:00", end: "2023-01-14T12:00:00+01:00", price: 336 },
+		{ start: "2023-01-14T12:00:00+01:00", end: "2023-01-14T13:00:00+01:00", price: 336 },
+		{ start: "2023-01-14T13:00:00+01:00", end: "2023-01-14T14:00:00+01:00", price: 336 },
+		{ start: "2023-01-14T14:00:00+01:00", end: "2023-01-14T15:00:00+01:00", price: 336 },
+		{ start: "2023-01-14T15:00:00+01:00", end: "2023-01-14T16:00:00+01:00", price: 336 },
+		{ start: "2023-01-14T16:00:00+01:00", end: "2023-01-14T17:00:00+01:00", price: 327 },
+		{ start: "2023-01-14T17:00:00+01:00", end: "2023-01-14T18:00:00+01:00", price: 345 },
+		{ start: "2023-01-14T18:00:00+01:00", end: "2023-01-14T19:00:00+01:00", price: 500 },
+		{ start: "2023-01-14T19:00:00+01:00", end: "2023-01-14T20:00:00+01:00", price: 491 },
+		{ start: "2023-01-14T20:00:00+01:00", end: "2023-01-14T21:00:00+01:00", price: 482 },
+		{ start: "2023-01-14T21:00:00+01:00", end: "2023-01-14T22:00:00+01:00", price: 473 },
+		{ start: "2023-01-14T22:00:00+01:00", end: "2023-01-14T23:00:00+01:00", price: 454 },
+		{ start: "2023-01-14T23:00:00+01:00", end: "2023-01-15T00:00:00+01:00", price: 436 },
+	],
+};
 
-const chargingSlots = [{ start: "2022-12-22T21:00:00Z", end: "2022-12-22T23:00:00Z" }];
+const plan = {
+	duration: 8695555555555,
+	plan: [
+		{
+			start: "2023-01-11T13:35:04.444444445+01:00",
+			end: "2023-01-11T14:00:00+01:00",
+			price: 336,
+		},
+		{
+			start: "2023-01-11T14:00:00+01:00",
+			end: "2023-01-11T15:00:00+01:00",
+			price: 336,
+		},
+		{
+			start: "2023-01-11T15:00:00+01:00",
+			end: "2023-01-11T16:00:00+01:00",
+			price: 336,
+		},
+	],
+	unit: "gCO2eq",
+	power: 11040,
+};
 </script>
 
 <template>
 	<Story title="TargetChargePlan">
 		<Variant title="standard">
-			<TargetChargePlan
-				:price-slots="priceSlots1"
-				:charging-slots="chargingSlots"
-				currency="EUR"
-			/>
+			<TargetChargePlan :tariff="tariff" :plan="plan" />
 		</Variant>
 	</Story>
 </template>
